+++ packages/cascader/src/style.tsx
@@ -1,6 +1,5 @@
 import { globalColor, illaPrefix } from "@illa-design/theme"
-import { SerializedStyles } from "@emotion/serialize"
-import { css } from "@emotion/react"
+import { css, SerializedStyles } from "@emotion/react"
 
 // popup
 export function applyPopupStyle(): SerializedStyles {