+++ packages/config-provider/src/index.ts
@@ -6,4 +6,6 @@ export * from "./config"
 // locale
 export * from "./locale/en-US"
 export * from "./locale/zh-CN"
+export * from "./locale/ja-JP"
+export * from "./locale/ko-KR"
 export * from "./locale/def"