+++ packages/pagination/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @illa-design/pagination
 
+## 1.0.10
+
+### Patch Changes
+
+- fix bugs
+
+- Updated dependencies []:
+  - @illa-design/config-provider@1.0.10
+  - @illa-design/icon@1.0.10
+  - @illa-design/input@1.0.10
+  - @illa-design/select@1.0.10
+  - @illa-design/system@1.0.10
+  - @illa-design/theme@1.0.10
+
+## 1.0.9
+
+### Patch Changes
+
+- fix bugs
+
+- fix bugs
+
+- Updated dependencies []:
+  - @illa-design/config-provider@1.0.9
+  - @illa-design/icon@1.0.9
+  - @illa-design/input@1.0.9
+  - @illa-design/select@1.0.9
+  - @illa-design/system@1.0.9
+  - @illa-design/theme@1.0.9
+
 ## 1.0.8
 
 ### Patch Changes
@@ -7,6 +37,7 @@
 - fix bugs
 
 - Updated dependencies []:
+  - @illa-design/config-provider@1.0.8
   - @illa-design/icon@1.0.8
   - @illa-design/input@1.0.8
   - @illa-design/select@1.0.8