+++ packages/popconfirm/CHANGELOG.md
@@ -1,5 +1,39 @@
 # @illa-design/popconfirm
 
+## 1.0.10
+
+### Patch Changes
+
+- fix bugs
+
+- Updated dependencies []:
+  - @illa-design/button@1.0.10
+  - @illa-design/config-provider@1.0.10
+  - @illa-design/icon@1.0.10
+  - @illa-design/space@1.0.10
+  - @illa-design/system@1.0.10
+  - @illa-design/theme@1.0.10
+  - @illa-design/trigger@1.0.10
+  - @illa-design/typography@1.0.10
+
+## 1.0.9
+
+### Patch Changes
+
+- fix bugs
+
+- fix bugs
+
+- Updated dependencies []:
+  - @illa-design/button@1.0.9
+  - @illa-design/config-provider@1.0.9
+  - @illa-design/icon@1.0.9
+  - @illa-design/space@1.0.9
+  - @illa-design/system@1.0.9
+  - @illa-design/theme@1.0.9
+  - @illa-design/trigger@1.0.9
+  - @illa-design/typography@1.0.9
+
 ## 1.0.8
 
 ### Patch Changes
@@ -9,6 +43,8 @@
 - Updated dependencies []:
   - @illa-design/button@1.0.8
   - @illa-design/config-provider@1.0.8
+  - @illa-design/icon@1.0.8
+  - @illa-design/space@1.0.8
   - @illa-design/system@1.0.8
   - @illa-design/theme@1.0.8
   - @illa-design/trigger@1.0.8