+++ packages/progress/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @illa-design/progress
 
+## 1.0.10
+
+### Patch Changes
+
+- fix bugs
+
+- Updated dependencies []:
+  - @illa-design/icon@1.0.10
+  - @illa-design/space@1.0.10
+  - @illa-design/system@1.0.10
+  - @illa-design/theme@1.0.10
+  - @illa-design/trigger@1.0.10
+
+## 1.0.9
+
+### Patch Changes
+
+- fix bugs
+
+- fix bugs
+
+- Updated dependencies []:
+  - @illa-design/icon@1.0.9
+  - @illa-design/space@1.0.9
+  - @illa-design/system@1.0.9
+  - @illa-design/theme@1.0.9
+  - @illa-design/trigger@1.0.9
+
 ## 1.0.8
 
 ### Patch Changes