+++ packages/time-picker/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @illa-design/time-picker
 
+## 1.0.10
+
+### Patch Changes
+
+- fix bugs
+
+- Updated dependencies []:
+  - @illa-design/button@1.0.10
+  - @illa-design/config-provider@1.0.10
+  - @illa-design/icon@1.0.10
+  - @illa-design/input@1.0.10
+  - @illa-design/system@1.0.10
+  - @illa-design/theme@1.0.10
+  - @illa-design/trigger@1.0.10
+
+## 1.0.9
+
+### Patch Changes
+
+- fix bugs
+
+- fix bugs
+
+- Updated dependencies []:
+  - @illa-design/button@1.0.9
+  - @illa-design/config-provider@1.0.9
+  - @illa-design/icon@1.0.9
+  - @illa-design/input@1.0.9
+  - @illa-design/system@1.0.9
+  - @illa-design/theme@1.0.9
+  - @illa-design/trigger@1.0.9
+
 ## 1.0.8
 
 ### Patch Changes
@@ -7,6 +39,8 @@
 - fix bugs
 
 - Updated dependencies []:
+  - @illa-design/button@1.0.8
+  - @illa-design/config-provider@1.0.8
   - @illa-design/icon@1.0.8
   - @illa-design/input@1.0.8
   - @illa-design/system@1.0.8